feat(873): 4th, solution from discuss, AC
Save input array A to a hash set, and find the next fibonacii like
element in the set as long as possible.

Time complexity is O(n^2logM), where M is the max(A).

The key point is "fib sequence grows exponentially", that means the
maximum length of one fib-like subsequence won't be larger than logM.

Do not always getting stuck on trying to solve such problem with dp
at the begining.

Yes, I'm practicing formatting commit message.