feat(740): 1st, recursive, TLE
Use recursive with preprocessing the input array into 2 separate
arrays that one store sorted exclusive elements and the other
store the numbers according to the same index.

Pick the i-th element and separate the array into a[0, i-1] or
a[0, i-2] depending on a[i-1] equals to a[i] -1 or not.

Same the right side.