feat(279): 3rd, mathematical, AC
Really fast solution based on two formulas:

1. Lagrange's four-square theorem, also know as Bachet's conjecture,
states that every natural number can be represented as the sum of
four integer squares.

2. Legendre's three-square theorem states that a natural number can
be represented as the sum of three squares of integers
n = x^2 + y^2 + z^2 if and only if n is not of the form
n = 4^a * (8*b + 7) for integers a and b.

Math is everything! :)