feat(525): 2nd, math from discuss, AC
Reference:
https://leetcode.com/problems/continuous-subarray-sum/discuss/99499:

We iterate through the input array exactly once, keeping track of the
running sum mod k of the elements in the process. If we find that a
running sum value at index j has been previously seen before in some
earlier index i, then we know that the sub-array (i, j] contains a
desired sum.