feat(300): 2nd, O(NlogN) time from discuss, AC
https://leetcode.com/problems/longest-increasing-subsequence/discuss/74824:

`tails` is an array storing the smallest tail of all increasing
subsequences with length `i+1` in `tails[i]`.

For example, say we have nums = [4,5,6,3], then all the avaiable
increasing subsequences are:
len = 1: [4], [5], [6], [3] => tails[1] = 3
len = 2: [4,5], [5,6] => tails[2] = 5
len = 3" [4,5,6] => tails[3] = 6

We can easily prove that tails is a increasing array. Therefore it is
possible to do a binary search in tails array to find the one needs
update.

Each time we only do one of the two:
1. if num is larger than all tails, append it, increase the len by 1.
2. if tails[i-1] < num < tails[i], update tails[i]

Doing so will maintain the tails invariant. The final answer is just the
len.

...

This is a great solution! Take me a lot time to understand. Should
practice this problem again and again.